LPS-158298 Move LengthField to common components
