+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/App.scss
@@ -22,7 +22,6 @@
 @import 'fragment-configuration-fields/CSSClassSelectorField';
 @import 'fragment-configuration-fields/CustomCSSField';
 @import 'fragment-configuration-fields/HideFromSearchField';
-@import 'fragment-configuration-fields/LengthField';
 @import 'fragment-configuration-fields/SelectField';
 @import 'fragment-configuration-fields/TargetCollectionsField';
