+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/common/components/LengthField.scss
@@ -1,4 +1,4 @@
-@import 'atlas-variables';
+@import 'cadmin-variables';
 
 html#{$cadmin-selector} .cadmin .page-editor__length-field {
 	&__button {
